refacto(installcmd): split code.
