+++ config/install_command.go
@@ -2,7 +2,7 @@ package config
 
 import (
 	"os"
-
+	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 )
 
@@ -21,6 +21,27 @@ func InitInstallCmdConfig(config *InstallCmdConfig) error {
 	config.GoPackageFullPath = viper.GetString("package")
 	
 	return nil
+}
+
+// UpdateConfigAfterInstalling set the new bundle to true in config after install
+func UpdateConfigAfterInstalling(name string) error {
+	workdir, err := os.Getwd()
+
+	if err != nil {
+		return err
+	}
+
+	viper.AddConfigPath(workdir)
+	viper.SetConfigName(".go-cli-config")
+	viper.Set("bundles." + name, true)
+	// If a config file is found, read it in.
+	if err := viper.ReadInConfig(); err == nil {
+		log.Info("Using config file : ", viper.ConfigFileUsed())
+		viper.WriteConfig()
+		return nil
+	}
+	
+	return err
 }
 
 type InstallCmdConfig struct {
